initial git commit of Safe Strings Project
The Safe Strings Project:
=========================

Safe Strings is a small C library that handles C strings in a safe way.
The functions of this library look at the necessary space for the
operations and try to reserve that space (with malloc or realloc). The
functions only begin working with their tasks when the memory allocation
is successful.

http://sakuranohana.org/projects/sstrings/index.html
http://freshmeat.net/projects/sstrings/